requirements_cuda118.txt
absl-py==2.3.1
actionlib==1.12.1
affine==2.3.1
aioboto3==10.4.0
aiobotocore==2.4.2
aiofiles==23.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.2.0
angles==1.9.12
antlr4-python3-runtime==4.8
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==2.4.3
bondpy==1.8.5
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
casadi==3.5.5
catkin==0.7.29
certifi==2022.9.14
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.0
contourpy==1.0.5
control==0.9.1
controller_manager==0.18.4
controller_manager_msgs==0.18.4
coverage==6.4.4
cryptography==39.0.1
cv_bridge==1.13.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distlib==0.3.6
docker==6.0.0
drone_wrapper==1.3.10
dynamic_reconfigure==1.6.5
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.0.0
fastjsonschema==2.16.2
filelock==3.8.0
Fiona==1.8.21
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2023.1.0
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
geopandas==0.12.2
google-auth==2.40.3
google-auth-oauthlib==0.4.6
greenlet==1.1.3
grpcio==1.43.0
grpcio-tools==1.43.0
guppy3==3.1.2
h11==0.13.0
hf-xet==1.1.9
huggingface-hub==0.15.0
hydra-core==1.1.0rc1
hypothesis==6.67.1
identify==2.5.5
idna==3.4
image_geometry==1.13.1
importlib-metadata==4.12.0
iniconfig==1.1.1
interactive-markers==1.11.5
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
joint_state_publisher==1.12.15
json5==0.9.10
jsonschema==4.18.0
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
kdl-parser-py==1.13.3
kiwisolver==1.4.4
laser_geometry==1.6.7
lightning-utilities==0.15.2
lit==15.0.7
llvmlite==0.43.0
lxml==4.9.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mavros==1.16.0
mdurl==0.1.2
message_filters==1.14.13
mistune==2.0.4
mock==4.0.3
moto==4.1.2
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.4
munch==2.5.0
natten==0.14.6
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.2.1
nodeenv==1.7.0
notebook==6.4.12
notebook-shim==0.1.0
numba==0.60.0
numpy==1.23.4
# Editable install with no version control (nuplan-devkit==1.2.2)
-e /home/robosense/LK/project_pluto/nuplan-devkit
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
omegaconf==2.1.0rc1
opencv-python==4.5.1.48
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-trees==0.6.9
pyarrow==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
Pygments==2.13.0
pyinstrument==4.3.0
pyogrio==0.4.1
pyparsing==3.0.9
pyproj==3.4.0
pyquaternion==0.9.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python_qt_binding==0.4.4
pytorch-lightning==2.0.1
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.3.2
QtPy==2.2.0
rasterio==1.3.2
ray==2.0.0
referencing==0.36.2
requests==2.28.1
requests-oauthlib==2.0.0
resource_retriever==1.12.7
responses==0.22.0
retry==0.9.2
rich==14.1.0
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rpds-py==0.27.1
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_controller_manager==0.18.4
rqt_dep==0.4.9
rqt_drone_teleop==1.3.10
rqt_ez_publisher==0.5.0
rqt_graph==0.4.11
rqt_ground_robot_teleop==1.3.10
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_joint_trajectory_controller==0.17.3
rqt_joint_trajectory_plot==0.0.5
rqt_launch==0.4.8
rqt_launchtree==0.2.0
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_multiplot==0.0.10
rqt_nav_view==0.5.7
rqt_play_motion_builder==1.0.2
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_py_trees==0.3.1
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_rotors==2.2.3
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_virtual_joy==0.1.2
rqt_web==0.4.8
rsa==4.9.1
Rtree==1.0.0
rviz==1.13.30
s3fs==2023.1.0
s3transfer==0.6.0
safetensors==0.6.2
scipy==1.9.1
selenium==4.4.3
Send2Trash==1.8.0
sensor-msgs==1.12.8
shapely==2.0.1
six==1.16.0
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.27
stack-data==0.5.0
sympy==1.13.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
testbook==0.4.2
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
timm==1.0.19
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
topic_tools==1.14.13
torchmetrics==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
triton==2.0.0
typer==0.6.1
types-toml==0.10.8.4
typing_extensions==4.15.0
ujson==5.5.0
unique_id==1.0.6
urdfdom-py==0.4.6
urllib3==1.26.12
virtualenv==20.21.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.3
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
xacro==1.13.20
xmltodict==0.13.0
yarl==1.8.2
zipp==3.8.1
